+++ chaosk8s/__init__.py
@@ -14,7 +14,7 @@ from logzero import logger
 
 
 __all__ = ["create_k8s_api_client", "discover", "__version__"]
-__version__ = '0.14.0'
+__version__ = '0.15.0'
 
 
 def has_local_config_file():
@@ -116,4 +116,5 @@ def load_exported_activities() -> List[DiscoveredActivities]:
     activities.extend(discover_probes("chaosk8s.probes"))
     activities.extend(discover_actions("chaosk8s.pod.actions"))
     activities.extend(discover_probes("chaosk8s.pod.probes"))
+    activities.extend(discover_actions("chaosk8s.node.actions"))
     return activities